refactor(confidential transfers): split runtime.go into 2 files
